# docs/swagger.yaml
basePath: /
definitions:
  admin.CreateTokenRequest:
    properties:
      expire_at:
        type: string
      name:
        example: Test Token
        type: string
      scope:
        enum:
        - upload
        - pickup
        - admin
        example: upload,pickup
        type: string
    required:
    - name
    type: object
  admin.CreateTokenResponse:
    properties:
      data:
        $ref: '#/definitions/model.APIToken'
      token:
        type: string
    type: object
...
  model.APIToken:
    properties:
      created_at:
        type: string
      expire_at:
        type: string
      id:
        type: integer
      last_used_at:
        type: string
      name:
        type: string
      revoked:
        type: boolean
      scope:
        type: string
    type: object
...
  model.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      msg:
        example: success
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 自托管的文件暂存柜后端系统 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 文件暂存柜 API
  version: "1.0"
paths:
  /api/admin/api-tokens:
    get:
      description: 获取系统中所有 API Token 的详细信息（不包含哈希）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.APIToken'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - AdminAuth: []
      summary: 获取 API Token 列表
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: 创建一个新的 API Token，返回原始 Token（仅显示一次）
      parameters:
      - description: Token 信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.CreateTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.CreateTokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - AdminAuth: []
      summary: 创建 API Token
      tags:
      - Admin
  /api/admin/api-tokens/{id}:
    delete:
      description: 根据 ID 永久删除 API Token
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - AdminAuth: []
      summary: 删除 API Token
      tags:
      - Admin
  /api/admin/api-tokens/{id}/recover:
    post:
      description: 将已撤销的 API Token 恢复为有效状态
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - AdminAuth: []
      summary: 恢复 API Token
      tags:
      - Admin
  /api/admin/api-tokens/{id}/revoke:
    post:
      description: 将 API Token 标记为已撤销，使其失效但保留记录
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - AdminAuth: []
      summary: 撤销 API Token
      tags:
      - Admin
...
securityDefinitions:
  APITokenAuth:
    description: Type "Bearer <API-Token>" to authenticate. Required scope depends
      on the endpoint.
    in: header
    name: Authorization
    type: apiKey
  AdminAuth:
    description: Type "Bearer <JWT-Token>" or "Bearer <API-Token>" to authenticate.
      API Token must have 'admin' scope.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"